dns_updater: move methods for updating the dns-servers out of api/views
